+++ src/components/date_picker/date_picker.js
@@ -104,7 +104,7 @@ function DatePicker(
         popper.style.visibility = "hidden";
       }
 
-      if (ref.current.validInputValue) {
+      if (ref.current.validInputValue !== undefined) {
         setStringDate(ref.current.validInputValue);
 
         ref.current.validInputValue = undefined;
@@ -152,9 +152,6 @@ function DatePicker(
   format = getFormat(onlyMonthPicker, onlyYearPicker, format);
 
   [calendar, locale] = check(calendar, locale);
-
-  minDate = minDate && toDateObject(minDate, calendar);
-  maxDate = maxDate && toDateObject(maxDate, calendar);
 
   useEffect(() => {
     function handleClickOutside(event) {
@@ -517,7 +514,10 @@ function DatePicker(
         ignoreList: JSON.parse(formattingIgnoreList),
       });
 
-      if ((!minDate || date > minDate) && (!maxDate || date < maxDate)) {
+      if (
+        (!minDate || date > toDateObject(minDate, calendar, format)) &&
+        (!maxDate || date < toDateObject(maxDate, calendar, format))
+      ) {
         handleChange(date);
         onPropsChange?.({ ...datePickerProps, value: date });
 
@@ -577,7 +577,9 @@ function DatePicker(
         .concat(date)
         .flat()
         .some(
-          (date) => (minDate && date < minDate) || (maxDate && date > maxDate)
+          (date) =>
+            (minDate && date < toDateObject(minDate, calendar, format)) ||
+            (maxDate && date > toDateObject(maxDate, calendar, format))
         )
     ) {
       ref.current.validInputValue = getInputValue(value || ref.current.date);
@@ -585,7 +587,7 @@ function DatePicker(
       return setStringDate(newValue);
     }
 
-    ref.current.validInputValue = undefined;
+    ref.current.validInputValue = strDate;
 
     const mustUpdateState = onChange?.(date, {
       validatedValue: strDate,