+++ src/shared/toDateObject.js
@@ -1,10 +1,10 @@
 import DateObject from "react-date-object";
 
-export default function toDateObject(date, calendar) {
+export default function toDateObject(date, calendar, format = "YYYY/MM/DD") {
   if (date instanceof DateObject) {
-    date.setCalendar(calendar);
+    date.set({ calendar, format });
   } else {
-    date = new DateObject({ date, calendar });
+    date = new DateObject({ date, calendar, format });
   }
 
   return date;