+++ changelog.md
@@ -1,6 +1,7 @@
 # 4.3.4
 
 - CalendarRef & DatePickerRef added to the types [#232](https://github.com/shahabyazdi/react-multi-date-picker/issues/232)
+- Fixed a bug in the hour input of the AnalogTimePicker plugin [#233](https://github.com/shahabyazdi/react-multi-date-picker/issues/233)
 
 # 4.3.3